Refactored File Structure
